% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_below_threshold}
\alias{check_below_threshold}
\title{Check if first and last pH points passed as crossing the threshold
Must have interpolated values for support}
\usage{
check_below_threshold(
  xtime,
  xph,
  ph_threshold = 5.4,
  start_time = 0,
  end_time = 50
)
}
\description{
Check if first and last pH points passed as crossing the threshold
Must have interpolated values for support
}
