% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_individual_plot}
\alias{fit_individual_plot}
\title{Plot Individual pH Profiles}
\usage{
fit_individual_plot(fit)
}
\arguments{
\item{fit}{nlmixr2 fit object containing the observed and predicted values.}
}
\value{
A ggplot2 object representing the individual pH profiles plot.
}
\description{
Generates a plot of individual pH profiles over time for each subject/group, including observed points and predicted lines from a fitted model.
}
\author{
Omar I. Elashkar
}
