% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_obs_vs_ipred_plot}
\alias{fit_obs_vs_ipred_plot}
\title{Plot Observed vs Predicted pH}
\usage{
fit_obs_vs_ipred_plot(fit)
}
\arguments{
\item{fit}{nlmixr2 fit object containing the observed and predicted values.}
}
\value{
A ggplot2 object representing the observed vs predicted pH plot.
}
\description{
Generates a plot of observed pH values versus predicted pH values from a fitted model, including a reference line for perfect predictions.
}
\author{
Omar I. Elashkar
}
