% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_metrics_tab}
\alias{format_metrics_tab}
\title{Format pH Metrics Table}
\usage{
format_metrics_tab(x)
}
\arguments{
\item{x}{Data frame containing pH metrics to format.}
}
\value{
A gt table object representing the formatted pH metrics table.
}
\description{
Formats a data frame containing pH metrics for presentation, including formatting numbers, adding labels, and organizing columns into spanners.
}
\author{
Omar I. Elashkar
}
