% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatorPH-package.R
\docType{package}
\name{GatorpH-package}
\alias{GatorpH}
\alias{GatorpH-package}
\title{GatorpH: Quantification of Oral pH Dynamics}
\description{
GatorpH
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OmarAshkar/GatorpH}
  \item \url{https://omarashkar.github.io/GatorpH/}
  \item Report bugs at \url{https://github.com/OmarAshkar/GatorpH/issues}
}

}
\keyword{internal}
