% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_auc_linear_down_log_up}
\alias{get_auc_linear_down_log_up}
\title{Calculate AUC
Linear to tmin, log after tmin}
\usage{
get_auc_linear_down_log_up(time, conc)
}
\description{
Calculate AUC
Linear to tmin, log after tmin
}
