% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kpd_mod2}
\alias{kpd_mod2}
\title{KPD Model with Linear Parameterization}
\usage{
kpd_mod2(
  edk50 = 10,
  kde = 0.1,
  kd = 0.5,
  ks = 3.5,
  gamma = 1,
  eta.edk50 = 0.1,
  eta.kde = 0.1,
  eta.kd = 0.1,
  eta.ks = 0.1,
  sigma_add = 0.01
)
}
\arguments{
\item{edk50}{Dose producing 50% of Emax (default is 10).}

\item{kde}{Elimination rate constant (1/h) from the virtual compartment KDE (default is 0.1).}

\item{kd}{Elimination rate constant (1/h) from the response compartment (default is 0.5).}

\item{ks}{Zero-order production rate constant (default is 3.5).}

\item{gamma}{Hill coefficient (default is 1).}

\item{eta.edk50}{Inter-individual variability (IIV) on edk50 (default is 0.1).}

\item{eta.kde}{IIV on kde (default is 0.1).}

\item{eta.kd}{IIV on kd (default is 0.1).}

\item{eta.ks}{IIV on ks (default is 0.1).}

\item{sigma_add}{Additive residual error (default is 0.01).}
}
\value{
An rxode2 model object representing the KPD model with linear parameterization.
}
\description{
KPD Model with Linear Parameterization
}
\author{
Omar I. Elashkar
}
