% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_direct_estimation}
\alias{run_direct_estimation}
\title{Run direct estimation for pH data}
\usage{
run_direct_estimation(
  x,
  ph_threshold = 5.4,
  time_start = 0,
  time_end = 50,
  method = "linear",
  add_support_points = FALSE,
  use_baseline = TRUE
)
}
\arguments{
\item{x}{Data frame containing pH data.}

\item{ph_threshold}{pH threshold for calculations.}

\item{time_start}{Start time for area and time under pH calculation.}

\item{time_end}{End time for area and time under pH calculation.}

\item{method}{Method for AUC calculation ("linear", "log", "linear_down_log_up").}
}
\value{
A data frame with subject and group pH parameters
}
\description{
Runs direct estimation calculations for pH data, including minimum pH, time to minimum pH, time under pH threshold, and AUC under pH threshold.
}
