% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_steph_curve}
\alias{simulate_steph_curve}
\title{Simulate Oral pH-Time Curve}
\usage{
simulate_steph_curve(
  model,
  time = seq(0, 50, by = 0.1),
  dose = 100,
  dose_time = 5,
  nsub = 1,
  baseline = 7,
  step = 0.1,
  group = "A",
  ignoreBSV = TRUE,
  ignoreRUV = TRUE,
  include_gamma = TRUE
)
}
\arguments{
\item{model}{rxode2 model object representing the PK-PD model.}

\item{time}{Vector of time points for simulation (default is seq(0, 50, by = 0.1)).}

\item{dose}{Dose amount (default is 100).}

\item{nsub}{Number of subjects to simulate (default is 1).}

\item{baseline}{Baseline pH value. Default is 7. Must be either 1 or same number as number of subjects.}

\item{step}{Time step for simulation (default is 0.1).}

\item{group}{Group identifier for the subjects (default is "A").}

\item{ignoreBSV}{Logical indicating whether to ignore between-subject variability (default is TRUE).}

\item{ignoreRUV}{Logical indicating whether to ignore residual unexplained variability (default is TRUE).}

\item{include_gamma}{logical indicating whether to include the gamma parameter in the simulation (default is TRUE).}

\item{baseline_time}{Time of baseline to adjust time points (default is -5).}
}
\value{
A data frame containing the simulated pH-time data.
}
\description{
Simulates oral pH-time curves based on a given PK-PD model.
}
\author{
Omar I. Elashkar
}
