% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.QuantPH}
\alias{summary.QuantPH}
\title{Summarize Direct Estimation Results}
\usage{
\method{summary}{QuantPH}(object, ...)
}
\arguments{
\item{object}{Data frame containing direct estimation results with class 'QuantPH'.}
}
\value{
A data frame with summary statistics for AUC, time under pH threshold, minimum pH, and time to minimum pH.
}
\description{
Provides summary statistics for direct estimation results.
}
\author{
Omar I. Elashkar
}
