% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_pH_curve}
\alias{fit_pH_curve}
\title{Fit pH Curve using NLME}
\usage{
fit_pH_curve(
  data,
  model,
  stratify = FALSE,
  estmethod = "focei",
  cov_params = c("kd", "kde", "edk50", "gamma"),
  cov_fixedeffects = c("t.kd", "t.kde", "t.edk50", "t.gamma"),
  include_gamma = TRUE
)
}
\arguments{
\item{data}{Data frame containing pH data. Must include columns: pH, time, id, group.}

\item{model}{rxode2/nlmixr2 model to fit.}

\item{stratify}{Logical indicating whether to fit separate models for each group (default is FALSE).}

\item{estmethod}{Estimation method to use (default is "focei").}

\item{include_gamma}{logical indicating whether to include the gamma parameter in the model (default is TRUE).}
}
\value{
nlmixr2 fit object.
}
\description{
Fits a pharmacodynamic model to pH data using nonlinear mixed-effects modeling. If `stratify` is TRUE, fits separate models for each group.
}
\author{
Omar I. Elashkar
}
