% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_pH_curve}
\alias{fit_pH_curve}
\title{Fit pH Curve using NLME}
\usage{
fit_pH_curve(
  data,
  model,
  amt,
  stratify = FALSE,
  estmethod = "focei",
  dose_time = 5
)
}
\arguments{
\item{data}{Data frame containing pH data. Must include columns: pH, time, id, group.}

\item{model}{rxode2/nlmixr2 model to fit.}

\item{amt}{Dose amount to administer at time 0.}

\item{stratify}{Logical indicating whether to fit separate models for each group (default is FALSE).}

\item{estmethod}{Estimation method to use (default is "focei").}

\item{dose_time}{Time (positive) of baseline to add before time 0 (default is 5).}
}
\value{
nlmixr2 fit object.
}
\description{
Fits a pharmacodynamic model to pH data using nonlinear mixed-effects modeling. If `stratify` is TRUE, fits separate models for each group.
}
\author{
Omar I. Elashkar
}
