% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integratepHArea}
\alias{integratepHArea}
\title{Integrate Area Under pH Threshold
Calculates the area under the pH curve below a specified pH threshold using the trapezoidal rule.
It assumes the flipped shape}
\usage{
integratepHArea(
  x,
  y,
  ph_threshold = 5.4,
  time_start = 0,
  time_end = 50,
  method = "linear",
  interpolate = TRUE,
  add_support_points = FALSE,
  plot = FALSE
)
}
\arguments{
\item{x}{Vector of time points.}

\item{y}{Vector of pH values corresponding to the time points.}

\item{ph_threshold}{pH threshold (default is 5.4).}

\item{time_start}{Start time for integration (default is 0).}

\item{time_end}{End time for integration (default is 50).}
}
\value{
Numeric value representing the area under the pH curve below the threshold.
}
\description{
Integrate Area Under pH Threshold
Calculates the area under the pH curve below a specified pH threshold using the trapezoidal rule.
It assumes the flipped shape
}
\author{
Omar I. Elashkar
}
