% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kpd_mod}
\alias{kpd_mod}
\title{KPD Model}
\usage{
kpd_mod(
  edk50 = 10,
  kde = 0.1,
  kd = 0.5,
  ks = 3.5,
  gamma = 1,
  eta.edk50 = 0.1,
  eta.kde = 0.1,
  eta.kd = 0.1,
  eta.ks = 0.1,
  sigma_add = 0.01
)
}
\arguments{
\item{edk50}{Numeric. Dose producing 50% of Emax (default is 10).}

\item{kde}{Numeric. Elimination rate constant (1/h) from the virtual compartment KDE (default is 0.1).}

\item{kd}{Numeric. First-order loss rate of response (default is 0.5).}

\item{ks}{Numeric. Maximum stimulatory effect on response (default is 3.5).}

\item{gamma}{Numeric. Hill coefficient (default is 1).}

\item{eta.edk50}{Numeric. Varianceof inter-individual variability for edk50 (default is 0.1).}

\item{eta.kde}{Numeric. Variance of inter-individual variability for kde (default is 0.1).}

\item{eta.kd}{Numeric. Variance of inter-individual variability for kd (default is 0.1).}

\item{eta.ks}{Numeric. Variance of inter-individual variability for ks (default is 0.1).}

\item{sigma_add}{Numeric. Additive error standard deviation for the response (default is 0.01).}
}
\value{
A rxode2 model object representing the KPD model.
}
\description{
A kinetic-pharmacodynamic (KPD) model for pH response to drug administration, based on a virtual compartment representing drug elimination and its effect on pH. The model includes parameters for the dose producing 50% of the maximum effect (EDK50), the elimination rate constant from the virtual compartment (KDE), the first-order loss rate of response (KD), the maximum stimulatory effect on response (KS), and the Hill coefficient (gamma). The model also accounts for inter-individual variability in these parameters and includes an additive error term for the response.
}
\author{
Omar I. Elashkar
}
