% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pHMetrics_from_fit}
\alias{pHMetrics_from_fit}
\title{Get pH Metrics from NLME Fit}
\usage{
pHMetrics_from_fit(
  x,
  ph_threshold = 5.4,
  time_start = 0,
  time_end = 50,
  step = 0.1,
  dose = 100,
  onlymean = TRUE,
  plot = FALSE,
  stratify_by = "None"
)
}
\arguments{
\item{x}{nlmixr2 fit object.}

\item{ph_threshold}{pH threshold for calculations (default is 5.4).}

\item{time_start}{Start time for simulation.}

\item{time_end}{End time for simulation.}

\item{step}{Time step for simulation.}

\item{dose}{Dose amount to administer at time 0.}

\item{onlymean}{Logical indicating whether to simulate only the mean profile (default is TRUE).}

\item{plot}{Logical indicating whether to plot the simulated pH profiles (default is FALSE).}

\item{stratify_by}{Stratification option for plotting ("None", "Subject", "Group").}
}
\value{
A data frame containing pH metrics calculated from the simulated pH profiles.
}
\description{
Get pH Metrics from NLME Fit
}
\author{
Omar I. Elashkar
}
