% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_pH_time}
\alias{plot_pH_time}
\title{Plot pH vs Time with Threshold}
\usage{
plot_pH_time(
  res,
  ph_threshold = 5.4,
  show_id = TRUE,
  stratify_by = "None",
  showAvg = FALSE
)
}
\arguments{
\item{res}{Data frame containing pH data.}

\item{ph_threshold}{pH threshold to highlight on the plot (default is 5.4).}
}
\value{
A ggplot2 object representing the pH vs time plot.
}
\description{
Generates a plot of pH values over time for different subjects/groups, highlighting a specified pH threshold.
}
\author{
Omar I. Elashkar
}
