% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_pH}
\alias{read_pH}
\title{Read pH Data}
\usage{
read_pH(file_path, dose_time = 1, amt = 100)
}
\arguments{
\item{file_path}{Path to the data file.}

\item{dose_time}{Time of dose administration (default is 1).}

\item{amt}{Dose amount (default is 100).}
}
\value{
A data frame containing the pH data.
}
\description{
Reads pH data from a CSV or Excel file and performs basic validation.
}
\author{
Omar I. Elashkar
}
